Small wrapper for SentencePiece tokenizer to avoid needing HF LlamaTokenizer in new model

+++ model.py
@@ -432,7 +432,7 @@ class ExLlama(nn.Module):
                 device = self.config.device_map.map(key)
                 tensor = f.get_tensor(key)
 
-                if key.endswith(".scales"): tensor = tensor.half().float()  # In reference this gets downcast, TODO: Test if ppl is better without this
+                if key.endswith(".scales"): tensor = tensor
 
                 tensor = tensor.to(device, non_blocking = True)
                 tensors[key] = tensor
